Fixes #8477. size_t is too small on some platforms to hold a filesize. off_t should work on any UNIX99 compatible system, but I went with uint64_t since it will always be sufficiently large and we're not using this in any posix calls.
git-svn-id: http://svn.mythtv.org/svn/trunk@24788 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/programs/mythbackend/upnpcdstv.cpp
@@ -10,6 +10,7 @@
 
 // POSIX headers
 #include <limits.h>
+#include <stdint.h>
 
 // MythTV headers
 #include "upnpcdstv.h"
@@ -252,7 +253,7 @@ void UPnpCDSTv::AddItem( const QString           &sObjectId,
     QString        sCategory    = query.value( 6).toString();
     QString        sHostName    = query.value( 7).toString();
     QString        sRecGroup    = query.value( 8).toString();
-    size_t         nFileSize    = query.value( 9).toULongLong();
+    uint64_t       nFileSize    = query.value( 9).toULongLong();
     QString        sBaseName    = query.value(10).toString();
 
     QDateTime      dtProgStart  = query.value(11).toDateTime();
